Do not build and run dev image
Travis CI has certain timeout limits (currently 50 min per build). We've already reached this. There is another limit of 10 min which stops the build if no output is produced. The R script in the dev image to install dev versions of the packages from github potentially clashes with this timeout as well.